backend/requirements.txt
pillow==10.4.0
tensorflow==2.17.0
fastapi[standard]==0.115.0
